refactor: RacingGame, RacingController 병합
